# Tidy up vm_push_frame