For instance its return value is used from nowhere so why not return `void`.

It seems compilers are smarter than me.  Neither speedup nor slowdown can be observed.  Must harm no one, though.

```
master: ruby 2.8.0dev (2020-07-01T13:41:16Z master 49029811d9) [aarch64-linux]
ours: ruby 2.8.0dev (2020-07-06T06:41:57Z origin/vm_push_frame 37bac56b10) [aarch64-linux]
Calculating -------------------------------------
                                       master                  ours
Optcarrot Lan_Master.nes   24.991672635908220    24.674321099849632 fps
                           25.049511891086389    24.751949030361153
                           25.107730926780810    24.832866393560870
                           25.193115588149983    25.084869359715185
                           25.227626632615966    25.130355271340051
                           25.247251028021367    25.223713670923104
                           25.323752498520427    25.294004296822507
```