@ko1 2020-07-09
No.

A little concern is performance on other compilers, CPU.
Clang on ARM machine?