@k0kubun 2020-07-09
Optcarrot seems fine at my environment (Linux x86_64, GCC 7.5.0, i7-4790K @ 4.00GHz) too:

```
$ benchmark-driver -v --rbenv 'before;after;before --jit;after --jit' benchmark.yml --repeat-count=12 --output=all
before: ruby 2.8.0dev (2020-07-05T16:15:47Z master efe851a0df) [x86_64-linux]
after: ruby 2.8.0dev (2020-07-07T00:02:22Z vm_push_frame 55a35649bd) [x86_64-linux]
before --jit: ruby 2.8.0dev (2020-07-05T16:15:47Z master efe851a0df) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-07-07T00:02:22Z vm_push_frame 55a35649bd) +JIT [x86_64-linux]
Calculating -------------------------------------
                                       before                 after          before --jit           after --jit
Optcarrot Lan_Master.nes    47.90696136067689     47.86544780120488     70.82988018325410     78.40795129070204 fps
                            48.08811127067737     48.41158522689150     76.12599867876445     84.58988633193086
                            49.37528171801291     48.96656206795582     80.94756718404652     84.93609714405245
                            49.60327985697335     49.40769262083241     84.74077401807905     85.37518847799471
                            50.06148885726757     49.45840926148377     88.83127270322659     87.59145996383153
                            50.24785115363551     49.66826263941514     91.22042530254973     87.77682286618247
                            50.45960005205968     49.75018717311129     91.70166656058082     88.48705192623078
                            50.57345777769526     49.81870364831044     91.81576041472785     90.54442643148857
                            50.64650459976789     49.84023408295305     91.83998096097912     91.05453996088721
                            50.74109511783863     50.10570633800229     91.93413327886388     91.23239918273043
                            50.75423127970495     50.50445419954861     91.99156212007854     91.62728127404294
                            50.82871791230230     50.93684142385430     92.89661780331625     92.00772730032472
```